feat(libraries): add meh_fork.h with improved simplified fork-join api
