+++ utils/libraries/SagiKimhi/meh/src/meh_darray.h
@@ -1,7 +1,7 @@
 #ifndef MEH_DARRAY_H_
 #define MEH_DARRAY_H_
 
-#include "meh_types.h"
+#include <sys/types.h>
 
 /* -----------------------------------------------------------------------------
  * Macros
@@ -30,64 +30,64 @@ struct meh_darray {
 # define meh_darray_create(T) meh_darray_create_(sizeof(T))
 #endif
 
-extern meh_darray_t
+extern struct meh_darray
 meh_darray_create_(size_t item_sz);
 
 extern void
-meh_darray_destroy(meh_darray_t *da);
+meh_darray_destroy(struct meh_darray *da);
 
 extern void *
-meh_darray_get(meh_darray_t da, size_t index);
+meh_darray_get(struct meh_darray da, size_t index);
 
 extern int
-meh_darray_set(meh_darray_t da, size_t index, const void *item);
+meh_darray_set(struct meh_darray da, size_t index, const void *item);
 
 extern void *
-meh_darray_pop(meh_darray_t *da);
+meh_darray_pop(struct meh_darray *da);
 
 extern void *
-meh_darray_last(meh_darray_t da);
+meh_darray_last(struct meh_darray da);
 
 extern void *
-meh_darray_first(meh_darray_t da);
+meh_darray_first(struct meh_darray da);
 
 extern int
-meh_darray_append(meh_darray_t *da, const void *item);
+meh_darray_append(struct meh_darray *da, const void *item);
 
 extern int
-meh_darray_extend(meh_darray_t *da, void **items, size_t nitems);
+meh_darray_extend(struct meh_darray *da, void **items, size_t nitems);
 
 extern size_t
-meh_darray_len(meh_darray_t da);
+meh_darray_len(struct meh_darray da);
 
 extern size_t
-meh_darray_blen(meh_darray_t da);
+meh_darray_blen(struct meh_darray da);
 
 extern int
-meh_darray_empty(meh_darray_t da);
+meh_darray_empty(struct meh_darray da);
 
 extern int
-meh_darray_non_empty(meh_darray_t da);
+meh_darray_non_empty(struct meh_darray da);
 
 extern int
-meh_darray_copy(meh_darray_t *dst, meh_darray_t src);
+meh_darray_copy(struct meh_darray *dst, struct meh_darray src);
 
 extern int
-meh_darray_take(meh_darray_t *dst, meh_darray_t *src);
+meh_darray_take(struct meh_darray *dst, struct meh_darray *src);
 
 extern int
-meh_darray_swap(meh_darray_t *dst, meh_darray_t *src);
+meh_darray_swap(struct meh_darray *dst, struct meh_darray *src);
 
 extern void
-meh_darray_print(meh_darray_t da, const char *name);
+meh_darray_print(struct meh_darray da, const char *name);
 
-extern meh_darray_t
-meh_darray_clone(meh_darray_t da);
-
-extern int
-meh_darray_resize(meh_darray_t *da, size_t newsz);
+extern struct meh_darray
+meh_darray_clone(struct meh_darray da);
 
 extern int
-meh_darray_resize_item(meh_darray_t *da, size_t newsz);
+meh_darray_resize(struct meh_darray *da, size_t newsz);
+
+extern int
+meh_darray_resize_item(struct meh_darray *da, size_t newsz);
 
 #endif
