+++ utils/libraries/SagiKimhi/meh/src/meh_misc.h
@@ -1,24 +1,24 @@
 #ifndef MEH_MISC_H_
 #define MEH_MISC_H_
 
-#ifndef meh_cast
-# define meh_cast(t, x) ((t)(x))
-#endif
-
-#ifndef meh_array_len
-# define meh_array_len(array) (sizeof(array) / sizeof((array)[0]))
-#endif
-
-#ifndef meh_dref_cast
-# define meh_dref_cast(t, x) meh_cast(t *, x)[0]
-#endif
-
 #ifndef meh_none
 # define meh_none meh_cast(void, 0)
 #endif
 
 #ifndef meh_as_none
 # define meh_as_none(x) meh_cast(void, (x))
+#endif
+
+#ifndef meh_cast
+# define meh_cast(t, x) ((t)(x))
+#endif
+
+#ifndef meh_dref_cast
+# define meh_dref_cast(t, x) meh_cast(t *, x)[0]
+#endif
+
+#ifndef meh_array_len
+# define meh_array_len(array) (sizeof(array) / sizeof((array)[0]))
 #endif
 
 #ifndef meh_swap
